Organize EndpointHandler to dispatch directive handlers
The EndpointHandler now receives a collection of directive handler
providers, and asks them to register handlers for it to dispatch on,
instead of hardcoding what methods it calls in the EndpointHandler
itself.

It also shims in a very rudimentary exception handler to send out an
Alexa.ErrorResponse on any runtime errors.